收集表1第一部分-插入代码修改

+++ src/main/java/com/hung/pojo/DemoCharacter.java
@@ -14,7 +14,7 @@ import lombok.Data;
 public class DemoCharacter {
     @TableId(type = IdType.AUTO)//id自增
     private Integer id;
-    private Integer subject_id;
+    private Long subject_id;
     private String name;
     private String gender;
     private String born_date;
